requirements.txt
Flask==3.1.1
pandas==2.3.0
pyodbc==5.2.0
python-dotenv==1.1.0
SQLAlchemy==2.0.41
